add dynmanic config system

+++ igen/Domain.h
@@ -98,7 +98,7 @@ public:
 
     void cleanup() override;
 
-    uint64_t config_space() const;
+    double config_space() const;
 
 public:
     vec <ptr<Config>> gen_all_configs() const;
